Squashed commit of the following:

commit d7633e38dde62c675be84fe6fb2427719bb6e36f
Merge: df1e381 2f8414e
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 15:01:16 2017 +0100

    Merge branch 'develop' into REL-170226_int_to_long

commit df1e381f537f51b4cca83872416985923fe1e403
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 14:29:48 2017 +0100

    Fix image file statistics unit tests

commit 48d5dd229701ca43c4de40cf2460c588d3653d47
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 13:36:28 2017 +0100

    Refactor also batching mechanism to use long instead of int.

commit 8bf68e31b5cc7e3b3f28926cfc8d6522185133dd
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 13:35:41 2017 +0100

    Change int to long in document statistics calculations.

    It is used by Total and FileSize statistics for Documents, Images and Native files. Please note that total documents count mechanism cannot be changed without modifications outside RIP (kCura.Relativity.Client.dll). Now we just case implicitly from int to long inside total count methods.

+++ Source/kCura.IntegrationPoints.Agent/Tasks/SyncManager.cs
@@ -237,7 +237,7 @@ namespace kCura.IntegrationPoints.Agent.Tasks
 			}
 		}
 
-		private void JobPostExecute(Job job, TaskResult taskResult, int items)
+		private void JobPostExecute(Job job, TaskResult taskResult, long items)
 		{
 			try
 			{
